+++ Backend/App/routes/augmentation_bp.py
@@ -1,7 +1,9 @@
 import logging
+from typing import List
 
 from flask import Blueprint, jsonify, request
 
+from Data.Files.StorageMethodology import StorageMethodology
 from Functionality.Augmentation import Augmentation
 from Utilities.AuthUtils import login_required
 from Utilities.Routing import parse_and_validate_data
@@ -13,6 +15,8 @@ AUGMENT_PROMPT_SCHEMA = {
 }
 QUESTION_PROMPT_SCHEMA = {
     "user_prompt": {"required": True, "type": str},
+    "selected_messages": {"required": False, "type": List},
+    "selected_files": {"required": False, "type": List},
 }
 
 
@@ -60,9 +64,29 @@ def question_user_prompt():
         data = request.get_json()
         parsed_data = parse_and_validate_data(data, QUESTION_PROMPT_SCHEMA)
         user_prompt = parsed_data.get("user_prompt")
+        selected_messages = parsed_data.get("selected_messages")
+        selected_files = parsed_data.get("selected_files")
+
+        reference_messages = None
+        if selected_messages:
+            reference_messages = []
+            for message in selected_messages:
+                reference_messages.append(
+                    message.get("prompt") + "\nResponse:\n" + message.get("response"))
+
+        reference_files = None
+        if selected_files:
+            reference_files = []
+            for file in selected_files:
+                file_category = file.get("category_id")
+                file_name = file.get("name")
+                full_path = str(file_category) + "/" + file_name
+
+                file_contents = StorageMethodology.select().read_file(full_path)
+                reference_files.append(f"<{file_name}>\n{file_contents}\n</{file_name}>")
 
         logging.debug(f"Generating questions against user prompt, data: {parsed_data}")
-        questions_for_prompt = Augmentation.question_user_prompt(user_prompt)
+        questions_for_prompt = Augmentation.question_user_prompt(user_prompt, reference_messages, reference_files)
         logging.info(f"questions for user prompt: \n{questions_for_prompt}")
 
         return jsonify({"message": questions_for_prompt})