+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- 
+- Question prompt will also reference files and/or messages added to the prompt (no more: "Can you tell me what language
+  (lorem.py) is written in...)
 
 ### Removed